requirements.txt
Babel==2.3.4
decorator==4.0.10
docutils==0.12
ebaysdk==2.1.4
feedparser==5.2.1
gevent==1.1.2
greenlet==0.4.10
jcconv==0.2.3
Jinja2==2.10.1
lxml==3.5.0
Mako==1.0.4
MarkupSafe==0.23
mock==2.0.0
ofxparse==0.16
passlib==1.6.5
Pillow==3.4.1
psutil==4.3.1
psycogreen==1.0
psycopg2==2.7.3.1
pydot==1.2.3
pyparsing==2.1.10
pyPdf==1.13
pyserial==3.1.1
Python-Chart==1.39
python-dateutil==2.5.
cryptography==2.7
python-openid==2.2.5
pytz==2016.7
pyusb==1.0.0
PyYAML==3.12
qrcode==5.3
reportlab==3.3.0
requests==2.20.0
six==1.10.0
suds-jurko==0.6
vatnumber==1.2
vobject==0.9.3
Werkzeug==0.11.11
wsgiref==0.1.2
XlsxWriter==0.9.3
xlwt==1.1.2
xlrd==1.0.0
